Fixed crash when drawing strings that terminate with \n Fixed many buttons having incorrect tint Increased version number to 4.0 Fixed sizing of characters in S'tack editor Fixed crash when rendering fusion plasma texture

+++ src/net/ncplanner/plannerator/planner/gui/menu/component/ScrollableCodeEditor.java
@@ -22,7 +22,7 @@ public class ScrollableCodeEditor extends Scrollable{
     public void render2d(double deltaTime){
         Renderer renderer = new Renderer();
         ArrayList<FormattedText> lines = editor.textDisplay.splitLines();
-        int xOff = (lines.size()+"").length()*editor.textWidth;
+        float xOff = (lines.size()+"").length()*editor.textWidth;
         renderer.setColor(Core.theme.getCodeLineMarkerColor());
         renderer.fillRect(x, y, x+xOff+editor.border*3/2, y+height);
         super.render2d(deltaTime);
